Move the MdnsService to a service module
